+Output Rectangle and (maybe?) Triangle

+++ karnauhova.alexandra/T1/shape.hpp
@@ -9,6 +9,6 @@ public:
   virtual rectangle_t getFrameRect() = 0;
   virtual void move(double x, double y) = 0;
   virtual void move(point_t t) = 0;
-  virtual void scale(point_t t, int k) = 0;
+  virtual void scale(point_t t, double k) = 0;
 };
 #endif
